Dependency link changes between BookStore.Services and BookStore.Infrastructure
