+++ project/BooksStore.Common/Exceptions/NotFoundException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace BooksStore.Common.Exceptions
+{
+    public class NotFoundException : Exception
+    {
+        public NotFoundException(string name, object key) : base($"Entity \"{name}\" ({key}) was not found.") { }
+    }
+}
