+++ project/BooksStore.Infrastructure/BooksStore.Infrastructure.csproj
@@ -18,6 +18,7 @@
   <ItemGroup>
     <ProjectReference Include="..\BooksStore.Common\BooksStore.Common.csproj" />
     <ProjectReference Include="..\BooksStore.Core\BooksStore.Core.csproj" />
+    <ProjectReference Include="..\BooksStore.Service\BooksStore.Services.csproj" />
   </ItemGroup>
 
 </Project>
