--- project/BooksStore.Infrastructure/Exceptions/NotFoundException.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace BooksStore.Common.Exceptions
-{
-    public class NotFoundException : Exception
-    {
-        public NotFoundException(string name, object key) : base($"Entity \"{name}\" ({key}) was not found.") { }
-    }
-}
