+++ project/BooksStore.Infrastructure/Implementation/Repositories/OrderRepository.cs
@@ -1,6 +1,6 @@
 ﻿using BooksStore.Core.Entities;
 using BooksStore.Infastructure.Data;
-using BooksStore.Infastructure.Interfaces.Repositories;
+using BooksStore.Services.Interfaces.Repositories;
 using Microsoft.EntityFrameworkCore;
 using QueryableFilterSpecification.Interfaces;
 using System.Collections.Generic;
