--- project/BooksStore.Infrastructure/Interfaces/Repositories/IDiscountRepository.cs
@@ -1,15 +1,0 @@
-﻿using BooksStore.Core.Entities;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace BooksStore.Infrastructure.Interfaces.Repositories
-{
-    public interface IDiscountRepository
-    {
-        Task AddAsync(Discount discount);
-        Task RemoveAsync(Discount discount);
-        Task UpdateAsync(Discount discount);
-        Task<IEnumerable<Discount>> GetAsync(int skip, int take);
-        Task<Discount> GetByBookIdAsync(int bookId);
-    }
-}
