+++ project/BooksStore.Service/BooksStore.Services.csproj
@@ -7,6 +7,7 @@
   <ItemGroup>
     <PackageReference Include="AutoMapper" Version="10.0.0" />
     <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="6.0.1" />
     <PackageReference Include="QueryableFilterSpecification" Version="1.0.1" />
     <PackageReference Include="System.Linq.Expressions" Version="4.3.0" />
   </ItemGroup>
@@ -14,7 +15,6 @@
   <ItemGroup>
     <ProjectReference Include="..\BooksStore.Common\BooksStore.Common.csproj" />
     <ProjectReference Include="..\BooksStore.Core\BooksStore.Core.csproj" />
-    <ProjectReference Include="..\BooksStore.Infrastructure\BooksStore.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
