+++ project/BooksStore.Service/CacheOptions/CacheTimes.cs
@@ -1,0 +1,23 @@
+﻿namespace BooksStore.Services.CacheOptions
+{
+    public class CacheTimes
+    {
+        public const int BooksCacheTime = 10;
+
+        public const int BookCacheTime = 5;
+
+        public const int BasketCacheTime = 10;
+
+        public const int OrdersCacheTime = 15;
+
+        public const int CategoriesCacheTime = 10;
+
+        public const int CategoryCacheTime = 15;
+
+        public const int CommentsCacheTime = 15;
+
+        public const int BooksByCategoryCacheTime = 10;
+
+        public const int AuthorCacheTime = 15;
+    }
+}
