+++ project/BooksStore.Service/Implementation/Services/Base/OrderService.cs
@@ -1,7 +1,6 @@
 ﻿using AutoMapper;
 using BooksStore.Common.Exceptions;
 using BooksStore.Core.Entities;
-using BooksStore.Infrastructure.Interfaces;
 using BooksStore.Services.DTO.Order;
 using BooksStore.Services.Implementation.Filters.OrderFilters;
 using BooksStore.Services.Interfaces.Services.Base;
@@ -9,6 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using BooksStore.Services.Interfaces.Repositories;
 
 namespace BooksStore.Services.Implementation.Services.Base
 {
