+++ project/BooksStore.Service/Implementation/Services/WithCaching/BasketCachingService.cs
@@ -1,14 +1,12 @@
 ﻿using AutoMapper;
 using BooksStore.Core.Entities;
-using BooksStore.Infrastructure.Interfaces;
+using BooksStore.Services.Interfaces;
 using BooksStore.Services.DTO.Basket;
 using BooksStore.Services.Interfaces.Services.Base;
 using BooksStore.Services.Interfaces.Services.WithCaching;
-using BooksStore.Web.CacheOptions;
 using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
+using BooksStore.Services.CacheOptions;
 
 namespace BooksStore.Services.Implementation.Services.WithCaching
 {
