+++ project/BooksStore.Service/Interfaces/ICacheManager.cs
@@ -1,0 +1,10 @@
+﻿namespace BooksStore.Services.Interfaces
+{
+    public interface ICacheManager
+    {
+        T Get<T>(string key);
+        bool IsSet(string key);
+        void Set<T>(string key, object data, int cacheTime);
+        void Remove(string key);
+    }
+}
