+++ project/BooksStore.Service/Interfaces/Repositories/IRepositoryFactory.cs
@@ -1,0 +1,12 @@
+﻿namespace BooksStore.Services.Interfaces.Repositories
+{
+    public interface IRepositoryFactory
+    {
+        public IBookRepository CreateBookRepository();
+        public IBasketRepository CreateBasketRepository();
+        public IAuthorRepository CreateAuthorRepository();
+        public ICategoryRepository CreateCategoryRepository();
+        public ICommentRepository CreateCommentRepository();
+        public IOrderRepository CreateOrderRepository();
+    }
+}
