+++ project/BooksStore.Service/Interfaces/Services/Base/IRoleService.cs
@@ -1,5 +1,5 @@
 ﻿using BooksStore.Common;
-using BooksStore.Infrastructure;
+using BooksStore.Services;
 using BooksStore.Services.DTO.Role;
 using System.Collections.Generic;
 using System.Threading.Tasks;
